Only cleanup if possible

+++ realm/realm-library/src/androidTestObjectServer/kotlin/io/realm/mongodb/MongoClientTest.kt
@@ -61,10 +61,11 @@ class MongoClientTest {
 
     @After
     fun tearDown() {
-        with(getCollectionInternal()) {
-            deleteMany(Document()).blockingGetResult()
+        if (this::client.isInitialized) {
+            with(getCollectionInternal()) {
+                deleteMany(Document()).blockingGetResult()
+            }
         }
-
         if (this::app.isInitialized) {
             app.close()
         }
